refactor: separate songs to map by hash
By using a map of each song by the song hash,
we are able to reuse the same data structure for both the tree
and the queue feature.

With this change, the list of songs in each directory,
is just a list of song hashes (still ordered by song name).

To retrieve any song in the tree, we'll need to use the `songsByHash`
map from here on.

We avoid saving the file handles in IndexedDB in order to save
bytes in the persisted database.

File system handles are serializable objects, which means that
the actual content of the file im bytes, would end up
in the database if left as part of each song.

This would cause overload the IndexedDB space, causing limits problem
(depending on vendor implementation), as well as occupying more space
in the user machine, as IndexedDB is persisted.

It'd almost be like all the songs exists twice in the user hard drive.