feat: save plyalist songs to firebase
In order to mirror the local data to Firebase database,
so that the public page can download the playlist songs,
we compress the data using gzip and then further splitting
into multiple chunks of 950 KB size each.

We do this so that we can save the whole tree of songs and
foldeers, without having to mirror the recursive structure
in Firestore.

This avoids us to hit the limits of writes and reads oer day,
as we don't add each song and dir as single documents of a
collection, but rather we save a bunch of byte blobs.

Firestore limit for each document is 1MB, so we can split our
big blob by chunks of 950KB each, to account for any space
used by firebase itself for each field.

Additionally, the gzip compression helps us save bytes on the
duplicated keywords in our data structure, as opposed to applying
any normalisation of the data upfront.